Navigation: move __experimental-fetch-link-suggestions to core-data
