+++ packages/core-data/src/fetch/__experimental-fetch-link-suggestions.js
@@ -1,0 +1,151 @@
+/**
+ * WordPress dependencies
+ */
+import apiFetch from '@wordpress/api-fetch';
+import { addQueryArgs } from '@wordpress/url';
+import { decodeEntities } from '@wordpress/html-entities';
+import { __ } from '@wordpress/i18n';
+
+/**
+ * Filters the search by type
+ *
+ * @typedef { 'post' | 'term' | 'post-format' } WPLinkSearchType
+ */
+
+/**
+ * @typedef WPLinkSearchOptions
+ *
+ * @property {boolean}             [isInitialSuggestions] Displays initial search suggestions, when true.
+ * @property {WPLinkSearchType}    [type]                 Filters by search type.
+ * @property {string}              [subtype]              Slug of the post-type or taxonomy.
+ * @property {number}              [page]                 Which page of results to return.
+ * @property {number}              [perPageArg]           Search results per page.
+ */
+
+/**
+ * @typedef WPLinkSearchResult
+ *
+ * @property {number} id     Post or term id.
+ * @property {string} url    Link url.
+ * @property {string} title  Title of the link.
+ * @property {string} type   The taxonomy or post type slug or type URL.
+ */
+
+/**
+ * Fetches link suggestions from the API.
+ *
+ * @async
+ * @typedef {( searchText: string, searchOptions: WPLinkSearchOptions ) => Promise< WPLinkSearchResult[] >} WPFetchLinkSuggestions
+ */
+
+/**
+ * Returns a function that when invoked, fetches link suggestions from the API.
+ *
+ * @param {Object}   [editorSettings]
+ * @param {boolean}  [editorSettings.disablePostFormats]
+ *
+ * @example
+ * ```js
+ * import { __experimentalFetchLinkSuggestions as createFetchLinkSuggestions } from '@wordpress/core-data';
+ *
+ * //...
+ *
+ * export function initialize( id, settings ) {
+ *
+ * settings.__experimentalFetchLinkSuggestions = createFetchLinkSuggestions(
+		settings
+	);
+ * ```
+ * @return { WPFetchLinkSuggestions } Function that fetches link suggestions
+ *
+ */
+const createFetchLinkSuggestions = ( {
+	disablePostFormats = false,
+} = {} ) => async ( search, searchOptions ) => {
+	const {
+		isInitialSuggestions = false,
+		type = undefined,
+		subtype = undefined,
+		page = undefined,
+		perPageArg = undefined,
+	} = searchOptions;
+	const perPage = perPageArg || isInitialSuggestions ? 3 : 20;
+
+	const queries = [];
+
+	if ( ! type || type === 'post' ) {
+		queries.push(
+			apiFetch( {
+				path: addQueryArgs( '/wp/v2/search', {
+					search,
+					page,
+					per_page: perPage,
+					type: 'post',
+					subtype,
+				} ),
+			} ).catch( () => [] ) // fail by returning no results
+		);
+	}
+
+	if ( ! type || type === 'term' ) {
+		queries.push(
+			apiFetch( {
+				path: addQueryArgs( '/wp/v2/search', {
+					search,
+					page,
+					per_page: perPage,
+					type: 'term',
+					subtype,
+				} ),
+			} ).catch( () => [] )
+		);
+	}
+
+	if ( ! disablePostFormats && ( ! type || type === 'post-format' ) ) {
+		queries.push(
+			apiFetch( {
+				path: addQueryArgs( '/wp/v2/search', {
+					search,
+					page,
+					per_page: perPage,
+					type: 'post-format',
+					subtype,
+				} ),
+			} ).catch( () => [] )
+		);
+	}
+
+	return Promise.all( queries ).then( ( results ) => {
+		return results
+			.reduce(
+				( accumulator, current ) => accumulator.concat( current ), //flatten list
+				[]
+			)
+			.filter(
+				/**
+				 * @param {{ id: number }} result
+				 */
+				( result ) => {
+					return !! result.id;
+				}
+			)
+			.slice( 0, perPage )
+			.map(
+				/**
+				 * @param {{ id: number, url:string, title?:string, subtype?: string, type?: string }} result
+				 */
+				( result ) => {
+					return {
+						id: result.id,
+						url: result.url,
+						title:
+							decodeEntities( result.title || '' ) ||
+							__( '(no title)' ),
+						type: result.subtype || result.type,
+					};
+				}
+			);
+	} );
+};
+
+export default createFetchLinkSuggestions;
