+++ packages/core-data/src/fetch/index.js
@@ -1,0 +1,1 @@
+export { default as __experimentalFetchLinkSuggestions } from './__experimental-fetch-link-suggestions';
