+++ packages/core-data/src/index.js
@@ -78,3 +78,4 @@ register( store );
 
 export { default as EntityProvider } from './entity-provider';
 export * from './entity-provider';
+export * from './fetch';
